Cleanup Makefile, .editorconfig, and whitespace (#69)
A section of the .travis.yml was indented with 4 spaces instead of the 2
specified for .yml files in the root .editorconfig.

Indentation of the docker-compose.yml has been fixed.

Trailing whitespace was removed from the README.md file.

The Makefile is now declares the added targets as PHONY.  The other
targets are now listed in the help and are also categorized.  The
wording of the help has been corrected.

The ui/.editorconfig is not the root.  Additions for other file formats
found in the ui/ folder have also been added.